inigo_py: split extract middlewares to subpackages
commit-id:d922910e
