+++ inigo_py/flask/__init__.py
@@ -1,0 +1,5 @@
+from .middleware import Middleware
+
+__all__ = [
+    'Middleware'
+]
